Now that the docker image is smaller, reenable caching.
The github action runner only guarantees 14GB, the ubuntu-latest image
has about 20GB.

Our compressed container image is about 7GB, which is exported when we
push, so worst case we have.

    7GB - previous (cache) image
    7GB - current image being rebuilt
    7GB - new layers getting exported to push

So it is obvious why we're running out of space in the runner. However
there are a lot of preinstalled tools we don't care about, Android SDKs,
dotnet environment, etc. So use an action that clears some of those
which should free up anywhere between 20 and 30 GB, giving us plenty of
space to build our image.